Relaxed parameter restrictions in IXArchADT to allow passing of "Serializable" rather than "String" values [1]  * TODO: Discuss and agree upon the best way forward with this issue (see [1])
[1] This change needs review. Assuming that XArchADT is on a server and being accessed from a client, the original code would rely on Java to translate a non-String value into a String (e.g., 1 -> "1", true -> "true", null -> "null", 0.1f -> "some arbitrary precision float string"). On the server side, XArchADTImpl would then "coerce" the String into the original value (e.g., "1" -> 1, ..., "some arbitrary prevision float string" -> 0.1f hopefully) and call the corresponding method on the underlying data model.

However, one of the models I'm using has XML:DateTime as a parameter (apparently, there are a lot of XML data types [2]). The EMF generated model requires a XMLCalendar [3] is supposed to be used to pass in an XML:DateTime. We could serialize this to a String, and then "coerce" the String back into a XMLCalendar. But, unfortunately, XMLCalendar looks to require additional inforation in it's constructor that is used to indicate the precision of the time (i.e., whether it's a date, a time of day on that date, or a time of day without a specific date). So, there's no class information on the server side to determine how to correctly coerce (read: in terms of precision) the XMLCalendar object from a String.

The solution in this commit is to pass Serializable objects, rather than just Strings. This still allows for distributed functionality. However, it also opens the door for someone passing a complex object (such as the aforementioned XMLCalendar) and then modifying it after it is embedded within XArchADT's model. This is, of course, bad. Perhaps we should look into cloning complex elements on the XArchADTImpl side instead of coercing objects. This way, we retain the type information, as well as the "not sharing pointers" property, but would additionally gain a performance improvement by not serializing (to String) and deserializing (from String) simple types.

[2] http://www.w3.org/TR/xmlschema-2/#built-in-datatypes

[3] http://download.eclipse.org/modeling/emf/emf/javadoc/2.6.0/org/eclipse/emf/ecore/xml/type/internal/XMLCalendar.html -- Note: this is an internal class, but it is exposed as part of the public API generated by EMF?!

git-svn-id: http://www.archstudio.org/svn/main/5/trunk@79 4f589ee1-8d18-4a0b-a29f-e26e9080690b