Fixed: XSDImporter would save files to the wrong location
XSDImporter seems to not support non-local files (e.g., projects imported from an external folder, linked files, etc.). This is a problem because, in such cases, the generated resources are saved in the wrong location. Later, when searching for the resources, they cannot be found and the process fails. To resolve this, we extend the base class and move the resources to the correct location before saving them.

git-svn-id: http://www.archstudio.org/svn/main/5/trunk@89 4f589ee1-8d18-4a0b-a29f-e26e9080690b